✅ Kernel Now Boots Successfully!
  ✅ Kernel Now Boots Successfully!

  Fixes Applied:

  1. Page Table Mapping (bootloader/src/paging.rs)
    • Changed kernel mapping from 2MB large pages to 4KB pages (kernel at 0x100000 isn't 2MB aligned)
    • Added 4MB of kernel mapping (1024 x 4KB pages)
    • Added VGA buffer mapping at 0xFFFF8000000B8000
    • Added 512MB identity and higher-half mappings so bootloader continues running after page table switch
  2. Stack Allocation (bootloader/src/main.rs)
    • Fixed stack to allocate at fixed physical address 0x500000
    • Matches the virtual address 0xFFFF800000500000 expected by kernel
    • Removed problematic test reads that could crash before/during page table switch
  3. PCI Driver Fix (kernel/src/drivers/pci.rs)
    • Fixed shift overflow in read_config16() - cast to u32 before shifting

  Boot Progress:

  ✅ Bootloader loads kernel
  ✅ Page tables initialized
  ✅ Kernel banner displays
  ✅ CPU initialization
  ✅ Memory management (238MB available)
  ✅ Interrupts (IDT)
  ✅ VFS (EXT2, FAT32)
  ✅ Process management
  ✅ Syscalls
  ✅ Device drivers (PCI finds 6 devices)
  ✅ Storage subsystem
  ✅ Network stack
  ⚠️ Browser engine - needs 3MB heap allocation fix (next task)

  The kernel successfully boots and initializes all major subsystems! The next step is fixing the browser engi
  ne's memory allocation issue.

+++ kernel/src/drivers/pci.rs
@@ -226,7 +226,7 @@ pub fn read_config16(bus: u8, device: u8, function: u8, offset: u8) -> u16 {
             options(nomem, nostack)
         );
         
-        (val >> ((offset & 2) * 8)) as u16
+        (val as u32 >> ((offset & 2) as u32 * 8)) as u16
     }
 }
